Create "INTERNPEAK_CS_04"
An IDS is a security technology designed to detect unauthorized access or anomalies within a network or system. It works by analyzing network traffic, system logs, and other data sources to identify suspicious or malicious behavior. When potential threats are detected, the IDS generates alerts or takes automated actions to mitigate the risks.